entity module: added utility.hpp, cleaned up fwd.hpp
