+++ src/entt/entity/group.hpp
@@ -9,6 +9,8 @@
 #include "../core/type_traits.hpp"
 #include "sparse_set.hpp"
 #include "storage.hpp"
+#include "utility.hpp"
+#include "entity.hpp"
 #include "fwd.hpp"
