+++ src/entt/entity/observer.hpp
@@ -11,6 +11,7 @@
 #include "../core/type_traits.hpp"
 #include "registry.hpp"
 #include "storage.hpp"
+#include "utility.hpp"
 #include "entity.hpp"
 #include "fwd.hpp"
